pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "tot-agent"
description = "Tree-of-Thought Location Agent — an inspectable, correctable memory engine that learns places through observation and correction."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Zachary Sturman" }]
keywords = ["agent", "memory", "perception", "location", "learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dynamic = ["version"]
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6.0",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "pip-audit>=2.7",
    "build>=1.2",
    "jsonschema>=4.21",
    "httpx>=0.27",
    "fastapi>=0.110",
]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
]
firestore = [
    "google-cloud-firestore>=2.16",
]

[project.scripts]
tot = "location_agent.cli:main"

[project.entry-points."tot.adapters"]
image = "location_agent.models:ImageAdapter"

[project.urls]
Homepage = "https://github.com/ZSturman/Train-of-Thought-Agent"
Repository = "https://github.com/ZSturman/Train-of-Thought-Agent"
Issues = "https://github.com/ZSturman/Train-of-Thought-Agent/issues"
Changelog = "https://github.com/ZSturman/Train-of-Thought-Agent/blob/main/CHANGELOG.md"

[tool.hatch.version]
path = "src/location_agent/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/location_agent"]

[tool.hatch.build.targets.sdist]
include = [
    "src/location_agent",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "pyproject.toml",
]
exclude = [
    "ARTICLE",
    "assets_for_notion",
    "runtime",
    "media/scenarios",
    "tests",
    "docs",
]

# ----------------------------------------------------------------------------
# pytest
# ----------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*", "*Tests"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]

[tool.coverage.run]
branch = true
source = ["location_agent"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false

# ----------------------------------------------------------------------------
# ruff
# ----------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]
extend-exclude = ["ARTICLE", "assets_for_notion", "runtime", "media/scenarios"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line too long — line-length already set
    "RUF005",  # iterable concatenation — stylistic; revisit in R2 cleanup
    "RUF059",  # unpacked-but-unused — common in test helpers
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ----------------------------------------------------------------------------
# mypy — strict on the public surface, lenient on the rest until R2
# ----------------------------------------------------------------------------
[tool.mypy]
python_version = "3.10"
files = ["src/location_agent"]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "location_agent.*"
disallow_untyped_defs = false
check_untyped_defs = true
